Connection: move the jetpack-ai-jwt REST route out of Agents Manager
My Jetpack required the whole Agents Manager package for this one class, so every plugin bundling My Jetpack shipped the AI sidebar without initializing it. The route only wraps Connection, so it now lives there; My Jetpack and Agents Manager register the Connection class, and the Agents Manager class stays as a deprecated alias.
