+++ projects/packages/connection/changelog/connect-437-ai-jwt-route
@@ -1,0 +1,4 @@
+Significance: minor
+Type: added
+
+REST API: add the `jetpack/v4/jetpack-ai-jwt` route, moved from the Agents Manager package.
