+++ projects/packages/my-jetpack/changelog/connect-437-ai-jwt-route
@@ -1,0 +1,4 @@
+Significance: patch
+Type: changed
+
+Register the `jetpack/v4/jetpack-ai-jwt` route from the Connection package and drop the Agents Manager dependency.
