+++ projects/plugins/protect/changelog/connect-437-ai-jwt-route
@@ -1,0 +1,5 @@
+Significance: patch
+Type: changed
+Comment: Update composer.lock.
+
+
